requirements.txt
coverage==7.6.12
et_xmlfile==2.0.0
iniconfig==2.0.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
tzdata==2025.1
